refactor v4 and v6 variants into separate files
